[fix] Move common DTO location
